pipelines/kafka_files/requirements_traffic.txt
kafka-python
requests
oauthlib
psycopg2
geopy
firebase-admin